# Inputs/openapi-specification.yaml
openapi: 3.0.3
info:
  title: Kurumsal Kripto Varlık Borsası API
  description: |
    **Techsonamy Kripto Exchange Platform API**
    
    Bu API, kurumsal kripto varlık borsası platformunun temel fonksiyonlarını sağlar:
    - Kullanıcı kimlik doğrulama ve yetkilendirme
    - Kripto varlık alım-satım işlemleri
    - Cüzdan yönetimi (yatırma/çekme)
    - Admin ve uyum fonksiyonları
    
    ## Önemli Notlar
    - Tüm decimal değerler **string** formatında gönderilmelidir (floating point hataları önlemek için)
    - Timestamp'ler **ISO 8601** formatında (UTC)
    - Rate limiting aktif: Header'larda `X-RateLimit-*` bilgileri döner
    - Hassas işlemler için 2FA gereklidir
    
    ## Güvenlik
    - OAuth 2.0 / JWT token bazlı kimlik doğrulama
    - HTTPS zorunlu (TLS 1.2+)
    - Request signing kritik endpoint'lerde gerekli
    
  version: 1.0.0
  contact:
    name: Techsonamy API Support
    email: api-support@techsonamy.com
    url: https://docs.techsonamy.com
  license:
    name: Proprietary
    url: https://techsonamy.com/license
servers:
  - url: https://api.example.com
    description: Production server (banka domain'i)
  - url: https://sandbox-api.example.com
    description: Sandbox environment (test)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/login endpoint
    
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for server-to-server communication
  schemas:
...
    SuccessResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        requestId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        data:
          type: object
        meta:
          type: object
          properties:
            page:
              type: integer
            pageSize:
              type: integer
            totalCount:
              type: integer
...
    Balance:
      type: object
      properties:
        asset:
          type: string
          example: BTC
        available:
          type: string
          example: "0.15000000"
          description: Available balance for trading/withdrawal
        locked:
          type: string
          example: "0.02000000"
          description: Locked in open orders
        total:
          type: string
          example: "0.17000000"
          description: Total = available + locked
        valueInTRY:
          type: string
          example: "314500.00"
...
paths:
...
  /api/v1/kyc/submit:
    post:
      tags:
        - users
      summary: Submit KYC documents
      description: Upload identity documents for KYC verification
      operationId: submitKYC
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - identityDocument
                - addressProof
                - selfiePhoto
              properties:
                identityDocument:
                  type: string
                  format: binary
                  description: ID card or passport image
                addressProof:
                  type: string
                  format: binary
                  description: Utility bill or bank statement
                selfiePhoto:
                  type: string
                  format: binary
                nationality:
                  type: string
                  example: TR
                taxId:
                  type: string
                  example: "12345678901"
      responses:
        '200':
          description: KYC submitted successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          kycId:
                            type: string
                          status:
                            type: string
                            enum: [PENDING_REVIEW]
                          submittedAt:
                            type: string
                            format: date-time
                          estimatedReviewTime:
                            type: string
                            example: "24-48 hours"

  # ==================== TRADING ENDPOINTS ====================
...
  /api/v1/wallets/balances:
    get:
      tags:
        - wallet
      summary: Get all balances
      description: Retrieve all asset balances for the user
      operationId: getBalances
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Balances
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Balance'
...